+++ test/tools/libtesteth/Options.cpp
@@ -53,7 +53,7 @@ void printHelp()
     cout << setw(30) << "--verbosity <level>" << setw(25) << "Set logs verbosity. 0 - silent, 1 - only errors, 2 - informative, >2 - detailed\n";
     cout << setw(30) << "--vm <name|path> (=legacy)" << setw(25) << "Set VM type for VMTests suite. Available options are: interpreter, legacy.\n";
     cout << setw(30) << "--evmc <option>=<value>" << setw(25) << "EVMC option\n";
-    cout << setw(30) << "--vmtrace" << setw(25) << "Enable VM trace for the test. (Require build with VMTRACE=1)\n";
+    cout << setw(30) << "--vmtrace" << setw(25) << "Enable VM trace for the test\n";
     cout << setw(30) << "--jsontrace <Options>" << setw(25) << "Enable VM trace to stdout in json format. Argument is a json config: '{ \"disableStorage\" : false, \"disableMemory\" : false, \"disableStack\" : false, \"fullStorage\" : true }'\n";
     cout << setw(30) << "--stats <OutFile>" << setw(25) << "Output debug stats to the file\n";
     cout << setw(30) << "--exectimelog" << setw(25) << "Output execution time for each test suite\n";
@@ -83,7 +83,7 @@ void printVersion()
 }
 }
 
-void Options::setVerbosity(int _level)
+void Options::setVerbosity(int _level, bool _vmTrace)
 {
     static boost::iostreams::stream<boost::iostreams::null_sink> nullOstream(
         (boost::iostreams::null_sink()));
@@ -98,6 +98,7 @@ void Options::setVerbosity(int _level)
     }
     else if (_level == 1 || _level == 2)
         logOptions.verbosity = VerbositySilent;
+    logOptions.vmTrace = _vmTrace;
     dev::setupLogging(logOptions);
 }
 
@@ -180,13 +181,8 @@ Options::Options(int argc, const char** argv)
         }
         else if (arg == "--vmtrace")
         {
-#if ETH_VMTRACE
             vmtrace = true;
             verbosity = VerbosityTrace;
-#else
-            cerr << "--vmtrace option requires a build with cmake -DVMTRACE=1\n";
-            exit(1);
-#endif
         }
         else if (arg == "--jsontrace")
         {
@@ -365,7 +361,7 @@ Options::Options(int argc, const char** argv)
     }
 
     // If no verbosity is set. use default
-    setVerbosity(verbosity == -1 ? 1 : verbosity);
+    setVerbosity(verbosity == -1 ? 1 : verbosity, vmtrace);
 }
 
 Options const& Options::get(int argc, const char** argv)